+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+### 2.0.3 (2016-07-04)
+
+- Added console command to get the status of the server
+  - ./bin/pjbservertools pjbserver:status <config_file>
+- Added console logger to all commands:
+  - To get more verbose message try the usual command with -v, -vv or -vvv
+    i.e: ./bin/pjbservertools -vvv pjbserver:start ./config/pjbserver.config.php.dist
+  
 ### 2.0.2 (2016-07-03)
 
 - Fix travis builds (no notable changes)
@@ -13,6 +21,10 @@
 - Refactored usage, config params must now be passed in the `StandaloneServer\Config($params)` object.
 - Support for PSR-3 logger interface.
 - Console commands added (symfony/console)
+  - ./bin/pjbservertools pjbserver:start <config_file>
+  - ./bin/pjbservertools pjbserver:stop <config_file>
+  - ./bin/pjbservertools pjbserver:restart <config_file>
+  
 
 
 ### 1.1.0 (2016-07-02)